Squashed 'examples/UnitUnified/NFCA/' changes from bd364f0..3fd067d
3fd067d Fixes bug and errors caused by the compiler environment

git-subtree-dir: examples/UnitUnified/NFCA
git-subtree-split: 3fd067d4074f7970181042c1140f0839b022704f

+++ Detect/main/Detect.cpp
@@ -119,7 +119,7 @@ void loop()
     if (nfc_a.detect(piccs)) {
         lcd.fillScreen(0);
         lcd.setCursor(0, 0);
-        uint32_t idx{};
+        uint16_t idx{};
         for (auto&& u : piccs) {
             M5.Speaker.tone(6000, 5);
             // detect only performs a provisional classification based on sak, so further identification is required
